refactor tests into separate client.test and error-cases.test fiels
